moved getTokenFromHeaders function
